rename and rexport state
